requirements.txt
requests
colorgram.py